Mejora en bloqueo de usuarios y eliminacion de eventos

+++ src/controllers/event/deleteEventController.js
@@ -1,18 +1,25 @@
 //Eve
-const { Event } = require("../../db");
+const { Event, Paystub } = require("../../db");
 
 const deleteEventController = async (eventId) => {
   const event = await Event.findByPk(eventId);
-
+  const pays = await Paystub.findAll({ where: { eventID: eventId } });
   if (!event) {
     throw new Error("Evento no encontrado");
   }
-  event.destroy({
-    where: {
-      eventID: eventId, // ID del evento que deseas eliminar lógicamente
-    },
-  });
-  return event;
+
+  if (pays) {
+    throw new Error(
+      "Este evento ya tiene pagos realizados por usuarios, por favor edite el evento para cancerlo"
+    );
+  } else {
+    event.destroy({
+      where: {
+        eventID: eventId,
+      },
+    });
+    return event;
+  }
 };
 
 module.exports = {
